Refactored NumberFieldUtils.toNumber as NumberUtils.parseNumber
